+++ lib/Doctrine/ODM/PHPCR/Query/Expression/Comparison.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Doctrine\ODM\PHPCR\Query\Expression;
+
+use Doctrine\Common\Collections\Expr\Comparison as BaseComparison;
+use Doctrine\Common\Collections\Expr\ExpressionVisitor;
+
+/**
+ * This class purpose is to provide  provide a place 
+ * for the LIKE constant. Everything  else is handled 
+ * in Doctrine\Common\Collections\Expr\Comparison
+ */
+class Comparison extends BaseComparison
+{
+    const LIKE = 'like';
+}