+++ tests/Doctrine/Tests/ODM/PHPCR/Functional/ReferenceTest.php
@@ -18,6 +18,8 @@ use Doctrine\Tests\Models\References\RefRefTestObj;
 use Doctrine\Tests\Models\References\RefTestPrivateObj;
 use Doctrine\Tests\Models\References\RefManyTestObj;
 use Doctrine\Tests\Models\References\RefManyTestObjForCascade;
+use Doctrine\Tests\Models\References\RefManyWithParentTestObjForCascade;
+use Doctrine\Tests\Models\References\ParentTestObj;
 
 use Doctrine\ODM\PHPCR\PHPCRException;
 
@@ -45,24 +47,16 @@ class ReferenceTest extends \Doctrine\Tests\ODM\PHPCR\PHPCRFunctionalTestCase
      */
     private $node;
 
-    private $referrerType;
-    private $referencedType;
-    private $referrerManyType;
-    private $referrerManyForCascadeType;
-    private $weakReferrerType;
-    private $hardReferrerType;
-    private $referrerDifType;
+    private $referrerType = 'Doctrine\Tests\Models\References\RefTestObj';
+    private $referencedType = 'Doctrine\Tests\Models\References\RefRefTestObj';
+    private $referrerManyType = 'Doctrine\Tests\Models\References\RefManyTestObj';
+    private $referrerManyForCascadeType = 'Doctrine\Tests\Models\References\RefManyTestObjForCascade';
+    private $hardReferrerType = 'Doctrine\Tests\Models\References\HardRefTestObj';
+    private $referrerDifType = 'Doctrine\Tests\Models\References\RefDifTestObj';
+    private $referrerManyWithParentForCascadeType = 'Doctrine\Tests\Models\References\RefManyWithParentTestObjForCascade';
 
     public function setUp()
     {
-        $this->referrerType = 'Doctrine\Tests\Models\References\RefTestObj';
-        $this->referencedType = 'Doctrine\Tests\Models\References\RefRefTestObj';
-        $this->referrerManyType = 'Doctrine\Tests\Models\References\RefManyTestObj';
-        $this->referrerManyForCascadeType = 'Doctrine\Tests\Models\References\RefManyTestObjForCascade';
-        $this->weakReferrerType = 'Doctrine\Tests\Models\References\WeakRefTestObj';
-        $this->hardReferrerType = 'Doctrine\Tests\Models\References\HardRefTestObj';
-        $this->referrerDifType = 'Doctrine\Tests\Models\References\RefDifTestObj';
-
         $this->dm = $this->createDocumentManager();
         $this->session = $this->dm->getPhpcrSession();
         $this->node = $this->resetFunctionalNode($this->dm);
@@ -1292,5 +1286,37 @@ class ReferenceTest extends \Doctrine\Tests\ODM\PHPCR\PHPCRFunctionalTestCase
             $i++;
         }
         $this->assertEquals(5, $i);
+    }
+
+    public function testManyCascadeWithParentDelete()
+    {
+        $refManyTestObjForCascade = new RefManyWithParentTestObjForCascade();
+        $refManyTestObjForCascade->id = "/functional/refManyWithParentTestObjForCascade";
+
+        $references = array();
+        for ($i = 0; $i < 3; $i++) {
+            $newRefCascadeManyTestObj = new ParentTestObj();
+            $newRefCascadeManyTestObj->nodename = "ref$i";
+            $newRefCascadeManyTestObj->name = "refCascadeWithParentManyTestObj$i";
+            $references[] = $newRefCascadeManyTestObj;
+        }
+        $refManyTestObjForCascade->setReferences($references);
+
+        $this->dm->persist($refManyTestObjForCascade);
+        $this->dm->flush();
+        $this->dm->clear();
+
+        $this->assertTrue($this->session->getNode("/functional")->hasNode("refManyWithParentTestObjForCascade"));
+
+        for ($i = 0; $i < 3; $i++) {
+            $this->assertTrue($this->session->getNode("/functional/refManyWithParentTestObjForCascade")->hasNode("ref$i"));
+        }
+
+        $referrer = $this->dm->find($this->referrerManyWithParentForCascadeType, '/functional/refManyWithParentTestObjForCascade');
+
+        $this->dm->remove($referrer);
+        $this->dm->flush();
+
+        $this->assertFalse($this->session->getNode("/functional")->hasNode("refManyWithParentTestObjForCascade"));
     }
 }