+++ tests/travis_doctrine_dbal.sh
@@ -18,5 +18,5 @@ fi
 mysql -e 'create database IF NOT EXISTS phpcr_odm_tests;' -u root
 
 cp ${SCRIPT_DIR}/../cli-config.doctrine_dbal.php.dist ${SCRIPT_DIR}/../cli-config.php
-${SCRIPT_DIR}/../bin/phpcr jackalope:init:dbal
-${SCRIPT_DIR}/../bin/phpcr doctrine:phpcr:register-system-node-types
+${SCRIPT_DIR}/../bin/phpcrodm jackalope:init:dbal
+${SCRIPT_DIR}/../bin/phpcrodm doctrine:phpcr:register-system-node-types