+++ tests/travis_jackrabbit.sh
@@ -18,4 +18,4 @@ fi
 ./vendor/jackalope/jackalope-jackrabbit/bin/jackrabbit.sh
 
 cp ${SCRIPT_DIR}/../cli-config.jackrabbit.php.dist ${SCRIPT_DIR}/../cli-config.php
-${SCRIPT_DIR}/../bin/phpcr doctrine:phpcr:register-system-node-types
+${SCRIPT_DIR}/../bin/phpcrodm doctrine:phpcr:register-system-node-types