+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+2013-01-18
+----------
+
+ * Removed DocumentRepository->getDocumentsByQuery as it is no longer needed.
+   createQuery returns you an ODM query that can find documents directly.
+   To get documents from a PHPCR query there is DocumentManager->getDocumentsByPhpcrQuery()
+
 2013-01-13
 ----------
 
@@ -12,4 +19,4 @@ Changelog
    ->execute(Query::HYDRATE_PHPCR);
  * CreateQuery($statement, $language) has NOT been implemented in the new query builder.
    It is, however, still available in the DocumentManager.
- * DocumetManager->getDocumentsByQuery renamed to getDocumentByPhpcrQuery()
+ * DocumetManager->getDocumentsByQuery renamed to getDocumentsByPhpcrQuery()